Start restructuring the tutorials.
